feat: game card animation
